feat(frontend): mysql、tendbcluster版本升级交互优化 #12934
