+++ dbm-ui/frontend/src/views/db-manage/mysql/MYSQL_LOCAL_UPGRADE/components/FixedModuleTargetVersionColumn.vue
@@ -25,53 +25,13 @@
         <div class="content-item">
           <div class="item-title">{{ t('绑定模块') }}：</div>
           <div class="item-content">
-            <TableEditSelect
-              ref="moduleSelectRef"
-              is-plain
-              :list="moduleSelectList"
-              :model-value="newDbModuleId"
-              :placeholder="t('请选择')"
-              :pop-width="240"
-              :rules="moduleRules"
-              @change="(value) => handleModuleChange(value as number)">
-              <template #default="{ item }">
-                <div class="module-option-item">
-                  <div class="module-option-label">
-                    {{ item.name }}
-                  </div>
-                  <div class="module-opiton-info">
-                    {{ item.info }}
-                  </div>
-                </div>
-              </template>
-              <template #footer>
-                <div class="module-select-footer">
-                  <BkButton
-                    class="plus-button"
-                    text
-                    @click="handleCreateModule">
-                    <DbIcon
-                      class="footer-icon mr-4"
-                      type="plus-8" />
-                    {{ t('跳转新建模块') }}
-                  </BkButton>
-                  <BkButton
-                    class="refresh-button"
-                    text
-                    @click="handleRefreshModule">
-                    <DbIcon
-                      class="footer-icon"
-                      type="refresh-2" />
-                  </BkButton>
-                </div>
-              </template>
-            </TableEditSelect>
+            {{ currentModule?.db_module_name || '' }}
           </div>
         </div>
         <div class="content-item">
           <div class="item-title">{{ t('数据库版本') }}：</div>
           <div class="item-content">
-            {{ currentModule?.spider_version || '' }}
+            {{ currentModule?.db_version || '' }}
           </div>
         </div>
         <div class="content-item">
@@ -121,31 +81,28 @@
   import { useI18n } from 'vue-i18n';
   import { useRequest } from 'vue-request';
 
-  import TendbClusterModel from '@services/model/tendbcluster/tendbcluster';
-  import { getSpiderVersionModules } from '@services/source/mysqlToolbox';
+  import TendbhaModel from '@services/model/mysql/tendbha';
+  import { getVersionModules } from '@services/source/mysqlToolbox';
 
-  import { TicketTypes } from '@common/const';
+  import { ClusterTypes } from '@common/const';
 
   import TableEditSelect from '@views/db-manage/mysql/common/edit/Select.vue';
 
-  type ModulesInfo = ServiceReturnType<typeof getSpiderVersionModules>[0];
+  type ModulesInfo = ServiceReturnType<typeof getVersionModules>[0];
 
   interface Props {
-    cluster: TendbClusterModel;
-    /**
-     * 高于当前集群主版本, 默认false
-     */
-    higherMajorVersion?: boolean;
-    /**
-     * 高于当前集群子版本, 默认false
-     */
-    higherSubVersion?: boolean;
+    cluster: {
+      cluster_type: ClusterTypes;
+      id: number;
+      master_domain: string;
+      related_clusters: {
+        id: number;
+        master_domain: string;
+      }[];
+    } & TendbhaModel;
   }
 
-  const props = withDefaults(defineProps<Props>(), {
-    higherMajorVersion: false,
-    higherSubVersion: false,
-  });
+  const props = defineProps<Props>();
 
   const modelValue = defineModel<{
     charset: string;
@@ -166,9 +123,6 @@
 
   const { t } = useI18n();
 
-  const route = useRoute();
-  const router = useRouter();
-
   const packageSelectList = ref<
     {
       id: number;
@@ -186,12 +140,6 @@
   >([]);
   const currentModule = ref<ModulesInfo>();
 
-  const moduleRules = [
-    {
-      message: t('绑定模块不能为空'),
-      validator: (value: string) => Boolean(value),
-    },
-  ];
   const packageRules = [
     {
       message: t('版本包文件不能为空'),
@@ -206,13 +154,13 @@
     },
   ];
 
-  const { run: fetchModuleList } = useRequest(getSpiderVersionModules, {
+  const { run: fetchModuleList } = useRequest(getVersionModules, {
     manual: true,
     onSuccess(data) {
       moduleSelectList.value = data.map((module) => ({
         ...module,
         id: module.db_module_id,
-        info: `${module.spider_version || ''}，${module.charset || ''}`,
+        info: `${module.db_version || ''}，${module.charset || ''}`,
         name: module.db_module_name,
       }));
 
@@ -247,8 +195,8 @@
       if (props.cluster.id) {
         fetchModuleList({
           cluster_id: props.cluster.id,
-          higher_major_version: props.higherMajorVersion,
-          higher_sub_version: props.higherSubVersion,
+          higher_major_version: false,
+          higher_sub_version: true,
         });
       }
     },
@@ -273,51 +221,6 @@
       pkgId.value = value;
       modelValue.value.pkg_name = findPackage.name;
     }
-  };
-
-  const handleModuleChange = (value: number) => {
-    newDbModuleId.value = value;
-    const findModule = moduleSelectList.value.find((item) => item.id === value) as unknown as ModulesInfo;
-    if (!findModule) return;
-    currentModule.value = findModule;
-    const pkgList = findModule.pkg_list.map((item) => ({
-      id: item.pkg_id,
-      name: item.pkg_name,
-    }));
-    packageSelectList.value = pkgList;
-    const [firstPackage] = pkgList;
-    if (firstPackage) {
-      pkgId.value = firstPackage.id;
-    }
-
-    modelValue.value = {
-      charset: findModule.charset,
-      db_module_name: findModule.db_module_name,
-      db_version: findModule.db_version,
-      pkg_name: firstPackage?.name || '',
-    };
-  };
-
-  const handleCreateModule = () => {
-    const url = router.resolve({
-      name: 'SelfServiceCreateDbModule',
-      params: {
-        bk_biz_id: window.PROJECT_CONFIG.BIZ_ID,
-        type: TicketTypes.TENDBCLUSTER_APPLY,
-      },
-      query: {
-        from: route.name as string,
-      },
-    });
-    window.open(url.href, '_blank');
-  };
-
-  const handleRefreshModule = () => {
-    fetchModuleList({
-      cluster_id: props.cluster.id,
-      higher_major_version: props.higherMajorVersion,
-      higher_sub_version: props.higherSubVersion,
-    });
   };
 </script>
 
@@ -353,46 +256,6 @@
       overflow: hidden;
       text-overflow: ellipsis;
       white-space: nowrap;
-    }
-  }
-
-  .module-option-item {
-    display: flex;
-    width: 100%;
-
-    .module-option-label {
-      flex: 1;
-      width: 0;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-
-    .module-opiton-info {
-      margin-left: auto;
-      color: #979ba5;
-    }
-  }
-
-  .module-select-footer {
-    display: flex;
-    height: 100%;
-    color: #63656e;
-    align-items: center;
-    justify-content: center;
-
-    .plus-button {
-      flex: 1;
-      padding-left: 36px;
-    }
-
-    .refresh-button {
-      width: 42px;
-      border-left: 1px solid #dcdee5;
-    }
-
-    .footer-icon {
-      font-size: 16px;
     }
   }
 </style>
